// src/app/service/rest/model/supplierUser.ts
/**
 * Water Solution Inventory Spring Boot App
 * Spring Boot REST API
 *
 * OpenAPI spec version: 1.0
 * Contact: watersolution@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Supplier } from './supplier';
import { User } from './user';


export interface SupplierUser { 
    responseCode?: string;
    responseDescription?: string;
    responseValues?: Array<string>;
    roleNameList?: Array<string>;
    supplier?: Supplier;
    user?: User;
}